requirements.txt
zhipuai
openai
python_snippets
click
streamlit
scipy
pandas
html2text
faiss-cpu
langchain
tiktoken
plotly
streamlit_chat
openpyxl